m10

+++ iCaRL_net.py
@@ -185,7 +185,7 @@ class iCaRL(nn.Module):
 
                 elif self.loss_config == 1:
                     #NLLL
-                    out = torch.log(torch.softmax(out, dim=1))
+                    out = torch.log(torch.softmax(out, dim=1)).long()
                     loss = self.clf_loss(torch.sigmoid(out[:, self.n_known:self.n_classes]), labels_hot[:, self.n_known:self.n_classes])
 
                 elif self.loss_config == 2:
